+++ cloud/connecting-algorithm/src/app.py
@@ -3,10 +3,10 @@ import os
 import boto3
 import requests
 import datetime
-from ortools.sat.python import cp_model
 
 import constants
 import database
+import match_logic
 
 # SET THIS TO YOUR LOCAL OPERATING SYSTEM
 LOCAL_OPERATING_SYSTEM = 'windows'  # set to 'mac', 'windows' or 'linux'
@@ -25,6 +25,7 @@ def lambda_handler(event, context):
         if len(users) % 2 == 1:
             users = [user for user in users if user['primaryKey']['S'] != ('USER#' + constants.JOKER_USER_ID)]
 
+        # Get previous match history for each user
         prev_matches = []
         for user in users:
             # Generate date of which last match will be considered
@@ -38,57 +39,7 @@ def lambda_handler(event, context):
             # Append to prev_matches a list of the IDs of the users the current user was matched with
             prev_matches.append([match['user2Id']['S'] for match in prev_matches_for_user])
 
-
-        # Creates the model
-        model = cp_model.CpModel()
-
-        # Creates the variables
-        x = {}
-        costs = {}
-
-        num_users = len(users)
-        for i in range(num_users):
-            for j in range(num_users):
-                if j <= i:
-                    x[i, j] = model.NewIntVar(0, 0, f'x[{i},{j}]')
-                    costs[i, j] = 0
-                else:
-                    x[i, j] = (model.NewBoolVar(f'x[{i},{j}]')) 
-                    costs[i, j] = eval_cost(users[i], users[j], prev_matches[i])
-
-        # Creates the constraints
-        for i in range(num_users):
-            model.Add(
-                sum((x[i, j] + x[j, i] for j in range(num_users))) == 1
-            )
-        
-        # Create objective function
-        objective = []
-        for i in range(num_users):
-            for j in range(num_users):
-                objective.append(costs[i, j] * x[i, j])
-        model.Minimize(sum(objective))
-
-        # Creates a solver and solves the model
-        solver = cp_model.CpSolver()
-        status = solver.Solve(model)
-
-        # Return array of all matches 
-        if status == cp_model.OPTIMAL or status == cp_model.FEASIBLE:
-            print('Solution found.')
-            print(f'Total cost = {solver.ObjectiveValue()}')
-            for i in range(num_users):
-                for j in range(i, num_users):
-                    if solver.BooleanValue(x[i, j]):
-                        new_match = {
-                            'user1Id': users[i]['primaryKey']['S'],
-                            'user1Faculty': users[i]['faculty']['S'],
-                            'user2Id': users[j]['primaryKey']['S'],
-                            'user2Faculty': users[j]['faculty']['S']
-                        }
-                        matches.append(new_match)
-        else:
-            print('No solution found.')
+        matches = match_logic.generate_matches(users, prev_matches)
 
     except requests.RequestException as e:
         print(e)
@@ -101,25 +52,6 @@ def lambda_handler(event, context):
             "message": matches,
         }),
     }
-
-
-# This will evaluate the potential 'cost' that a certain match between user1 and user2 will incur
-def eval_cost(user1, user2, prev_matches):
-    cost = 0
-
-    # If both users are from the same faculty
-    if user1['faculty']['S'] == user2['faculty']['S']:
-        cost += constants.COST_SAME_FACULTY
-
-    if user2['primaryKey']['S'] not in prev_matches: 
-        # If users have not been matched within the MATCH_COOLDOWN period
-        pass
-    else:
-        # If users have been matched within the MATCH_COOLDOWN period, add a cost that corresponds
-        # to the recency of the match - i.e. maximum if both users' last match was each other
-        cost += constants.COST_PREV_MATCHED[prev_matches.index(user2['primaryKey']['S'])]
-
-    return cost
 
 def get_client():
     isLocalEnvironment = os.environ['ENVIRONMENT'] == 'local'
